Add files via upload
Developed a responsive and user-friendly login form using HTML for structure and CSS for styling. The project involved creating a clean and modern interface with input fields for username and password, a "Remember Me" checkbox, and a "Forgot Password" link. Emphasized accessibility and cross-browser compatibility, ensuring the form works seamlessly across various devices and screen sizes. Utilized CSS Flexbox for layout alignment and implemented hover and focus effects to enhance user interaction. This project demonstrates proficiency in front-end development and attention to detail in UI design.